requirements.txt
torch
torchvision
pandas
numpy
Pillow
tensorboard
scikit-learn
matplotlib
ipykernel
onnxruntime